refactor: move Open Graph main types into directory
