+++ projects/ngx-meta/src/open-graph/index.ts
@@ -1,3 +1,3 @@
-export { OpenGraphMetadata } from './src/open-graph-metadata'
 export * from './src/basic-optional'
 export * from './src/profile'
+export * from './src/types'
