+++ projects/ngx-meta/src/open-graph/src/types/index.ts
@@ -1,0 +1,2 @@
+export { OpenGraph } from './open-graph'
+export { OpenGraphMetadata } from './open-graph-metadata'
