+++ projects/ngx-meta/src/open-graph/src/types/open-graph-metadata.ts
@@ -1,4 +1,4 @@
-import { OpenGraph } from './basic-optional'
+import { OpenGraph } from './open-graph'
 
 /**
  * Utility type to provide specific
