+++ projects/ngx-meta/src/open-graph/src/types/open-graph.ts
@@ -1,6 +1,5 @@
-import { OpenGraphImage } from './open-graph-image'
-import { OpenGraphType } from './open-graph-type'
-import { OpenGraphProfile } from '../../profile'
+import { OpenGraphImage, OpenGraphType } from '../basic-optional'
+import { OpenGraphProfile } from '../profile'
 
 /**
  * {@link https://ngx-meta.dev/built-in-modules/open-graph/ | Open Graph module}
